Split out resume tests from sign out tests
