+++ tests/integration/questionnaire/test_questionnaire_resume.py
@@ -1,24 +1,8 @@
 from tests.integration.integration_test_case import IntegrationTestCase
 
 
-class TestSignOut(IntegrationTestCase):
-    def test_sign_out_not_signed_in_redirects_to_signed_out_page(self):
-        self.get("/sign-out")
-        self.assertInUrl("/signed-out")
-
-    def test_sign_out_without_log_out_url_redirects_to_signed_out_page(self):
-        self.launchSurvey("test_textfield")
-        self.get("/sign-out")
-        self.assertInUrl("/signed-out")
-
-    def test_sign_out_with_log_out_url_redirects_to_log_out_url(self):
-        self.launchSurvey(
-            "test_textfield", account_service_log_out_url="https://localhost/logout"
-        )
-        self.get("/sign-out")
-        self.assertInUrl("/logout")
-
-    def test_sign_out_after_navigating_backwards(self):
+class TestResume(IntegrationTestCase):
+    def test_navigating_backwards(self):
         # If a user completes a block, then goes back and signs out, on re-authentication
         # they should resume on the first incomplete location
         self.launchSurvey("test_textfield")
